Further abstracted webhook functionality into plugin archetecture
