+++ lib/flint.js
@@ -39,7 +39,7 @@ class Flint extends EventEmitter {
     this.lexicon = [];
 
     // validate required options
-    if (!this.config.token || !this.config.webhookUrl) {
+    if (!this.config.token) {
       throw new Error('invalid or missing options');
     }
 
@@ -60,8 +60,6 @@ class Flint extends EventEmitter {
     // default status on constructor instantiation
     this.active = false;
     this.initialized = false;
-    this.isBotAccount = false;
-    this.isUserAccount = false;
 
     // initalize default plugins
     this.use('storage', MemStorage);
@@ -141,47 +139,17 @@ class Flint extends EventEmitter {
   // initialize flint
   initialize() {
     if (!this.initialized) {
-      let newWebhook = {};
       return this.spark.personMe()
         .then((person) => {
           const botPerson = _.cloneDeep(person);
-          botPerson.emails = [_.toLower(botPerson.emails[0])];
           if (botPerson.type === 'bot') {
-            this.isBotAccount = true;
-            this.isUserAccount = false;
-          } else {
-            this.isBotAccount = false;
-            this.isUserAccount = true;
+            botPerson.emails = [_.toLower(botPerson.emails[0])];
+            this.person = botPerson;
+            return when(true);
           }
-          return when(botPerson);
+          return when.reject(new Error('spark token is not for a bot account'));
         })
-        .then((botPerson) => {
-          this.person = botPerson;
-          newWebhook = {
-            name: `flint:${_.toLower(botPerson.emails[0])}`,
-            targetUrl: this.config.webhookUrl,
-            resource: 'all',
-            event: 'all',
-          };
-          return when(botPerson);
-        })
-        .then((botPerson) => {
-          const webhookName = newWebhook.name;
-          return this.spark.webhooksGet()
-            .then((webhooks) => {
-              const foundWebhooks = _.filter(webhooks, { name: webhookName });
-              if (foundWebhooks && foundWebhooks instanceof Array && foundWebhooks.length > 0) {
-                return when.map(webhooks, webhook => this.spark.webhookRemove(webhook.id))
-                  .catch((err) => {
-                    // log error but ignore
-                    this.logger.error(new Error(`could not remove all webhooks with name: ${webhookName}`));
-                    return when(true);
-                  });
-              }
-              return when(true);
-            });
-        })
-        .then(() => this.spark.webhookAdd(newWebhook))
+        .then(() => this.listener.init())
         .then(() => {
           this.emit('initialized');
           return when(true);
@@ -213,16 +181,6 @@ class Flint extends EventEmitter {
     } else {
       throw new Error('invalid plugin');
     }
-  }
-
-  // set spark token and reinit Spark
-  setSparkToken(token) {
-    if (token && typeof token === 'string') {
-      this.config.token = token;
-      this.spark = new Spark(this.config);
-      return when(this.config.token);
-    }
-    return when.reject(new Error('invalid token'));
   }
 
   // returns a bot object based on a query
